+++ packages/api-client/src/v2/blocks/operation-block/index.ts
@@ -1,3 +1,4 @@
-export { getSecuritySchemes } from './helpers/build-request-security'
+export { buildRequestSecurity, getSecuritySchemes } from './helpers/build-request-security'
 export { getExample } from './helpers/get-example'
+export { getResolvedUrl } from './helpers/get-resolved-url'
 export { default as OperationBlock } from './OperationBlock.vue'