+++ integrations/dotnet/aspnetcore/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @scalar/aspnetcore
 
+## 2.12.32
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/api-reference@1.44.11**
+
+## 2.12.31
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/api-reference@1.44.10**
+  - [#8016](https://github.com/scalar/scalar/pull/8016): feat: move history and auth into their own store
+
+- **@scalar/dotnet-shared@0.1.1**
+
 ## 2.12.30
 
 ### Patch Changes