+++ packages/json-magic/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @scalar/json-magic
 
+## 0.9.6
+
+### Patch Changes
+
+#### Updated Dependencies
+
+- **@scalar/helpers@0.2.11**
+  - [#8016](https://github.com/scalar/scalar/pull/8016): feat: move history and auth into their own store
+
 ## 0.9.5
 
 ### Patch Changes